+++ rpcs3/emucore.vcxproj.filters
@@ -992,6 +992,9 @@
     <ClCompile Include="Emu\RSX\D3D12\D3D12Texture.cpp">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\RSX\D3D12\D3D12Utils.cpp">
+      <Filter>Emu\GPU\RSX\D3D12</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Crypto\aes.h">
